update ImageUtils and ImagePaletteUtils acc to the reviews left
