--- shared/src/androidMain/kotlin/org/listenbrainz/shared/util/ImagePalette.android.kt
@@ -1,42 +1,0 @@
-package org.listenbrainz.shared.util
-
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.ImageBitmap
-import androidx.compose.ui.graphics.asAndroidBitmap
-import androidx.palette.graphics.Palette
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.withContext
-
-actual suspend fun getPaletteFromImage(bitmap: ImageBitmap): ImagePalette {
-    return withContext(Dispatchers.IO) {
-        val androidBitmap = bitmap.asAndroidBitmap()
-        val palette = Palette.from(androidBitmap).generate()
-        val lightColor =
-            palette.vibrantSwatch?.rgb ?: palette.mutedSwatch?.rgb ?: 0xFF888888.toInt()
-        val darkColor = palette.darkVibrantSwatch?.rgb ?: palette.darkMutedSwatch?.rgb ?: lightColor
-        val textColorDark =
-            palette.darkVibrantSwatch?.titleTextColor ?: palette.darkMutedSwatch?.titleTextColor
-            ?: 0xFFFFFFFF.toInt()
-        val bodyTextColorDark =
-            palette.darkVibrantSwatch?.bodyTextColor ?: palette.darkMutedSwatch?.bodyTextColor
-            ?: 0xFF000000.toInt()
-        val textColorLight =
-            palette.lightVibrantSwatch?.titleTextColor ?: palette.lightMutedSwatch?.titleTextColor
-            ?: 0xFFFFFFFF.toInt()
-        val bodyTextColorLight =
-            palette.lightVibrantSwatch?.bodyTextColor ?: palette.lightMutedSwatch?.bodyTextColor
-            ?: 0xFF000000.toInt()
-        ImagePalette(
-            gradientColors = listOf(
-                Color(lightColor),
-                Color(darkColor)
-            ),
-            titleColorLight = Color(textColorLight),
-            bodyTextColorLight = Color(bodyTextColorLight),
-            titleTextColorDark = Color(textColorDark),
-            bodyTextColorDark = Color(bodyTextColorDark),
-            lightBackgroundColor = Color(lightColor),
-            darkBackgroundColor = Color(darkColor)
-        )
-    }
-}
