+++ shared/src/commonMain/kotlin/org/listenbrainz/shared/util/ImagePaletteUtils.kt
@@ -2,8 +2,10 @@ package org.listenbrainz.shared.util
 
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.ImageBitmap
+import coil3.Bitmap
+import com.kmpalette.from
+import com.kmpalette.palette.graphics.Palette
 import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.IO
 import kotlinx.coroutines.withContext
 
 data class ImagePalette(
@@ -16,4 +18,35 @@ data class ImagePalette(
     val lightBackgroundColor: Color
 )
 
-expect suspend fun getPaletteFromImage(bitmap: ImageBitmap): ImagePalette
+suspend fun getPaletteFromImage(bitmap: ImageBitmap): ImagePalette {
+    return withContext(Dispatchers.Default) {
+        val palette = Palette.from(bitmap).generate()
+        val lightColor =
+            palette.vibrantSwatch?.rgb ?: palette.mutedSwatch?.rgb ?: 0xFF888888.toInt()
+        val darkColor = palette.darkVibrantSwatch?.rgb ?: palette.darkMutedSwatch?.rgb ?: lightColor
+        val textColorDark =
+            palette.darkVibrantSwatch?.titleTextColor ?: palette.darkMutedSwatch?.titleTextColor
+            ?: 0xFFFFFFFF.toInt()
+        val bodyTextColorDark =
+            palette.darkVibrantSwatch?.bodyTextColor ?: palette.darkMutedSwatch?.bodyTextColor
+            ?: 0xFF000000.toInt()
+        val textColorLight =
+            palette.lightVibrantSwatch?.titleTextColor ?: palette.lightMutedSwatch?.titleTextColor
+            ?: 0xFFFFFFFF.toInt()
+        val bodyTextColorLight =
+            palette.lightVibrantSwatch?.bodyTextColor ?: palette.lightMutedSwatch?.bodyTextColor
+            ?: 0xFF000000.toInt()
+        ImagePalette(
+            gradientColors = listOf(
+                Color(lightColor),
+                Color(darkColor)
+            ),
+            titleColorLight = Color(textColorLight),
+            bodyTextColorLight = Color(bodyTextColorLight),
+            titleTextColorDark = Color(textColorDark),
+            bodyTextColorDark = Color(bodyTextColorDark),
+            lightBackgroundColor = Color(lightColor),
+            darkBackgroundColor = Color(darkColor)
+        )
+    }
+}
