+++ shared/src/iosMain/kotlin/org/listenbrainz/shared/util/ImageUtils.ios.kt
@@ -2,26 +2,11 @@ package org.listenbrainz.shared.util
 
 import androidx.compose.ui.graphics.ImageBitmap
 import androidx.compose.ui.graphics.asComposeImageBitmap
-import coil3.SingletonImageLoader
+import coil3.Bitmap
 import coil3.request.ImageRequest
-import coil3.request.SuccessResult
-import coil3.toBitmap
-import org.listenbrainz.shared.repository.PlatformContext
 
-actual suspend fun fetchBitmapFromUrl(context: PlatformContext, url: String): ImageBitmap? {
-    return try {
-        val request = ImageRequest.Builder(coil3.PlatformContext.INSTANCE)
-            .data(url)
-            .build()
+actual fun getCoilPlatformContext(): coil3.PlatformContext = coil3.PlatformContext.INSTANCE
 
-        val result = SingletonImageLoader.get(coil3.PlatformContext.INSTANCE).execute(request)
-        if(result is SuccessResult){
-            result.image.toBitmap().asComposeImageBitmap()
-        } else{
-            null
-        }
-    } catch (e: Exception){
-        Log.e("Error while fetching bitmap: ${e.message}")
-        null
-    }
-}
+actual fun ImageRequest.Builder.platformConfig(): ImageRequest.Builder = this
+
+actual fun Bitmap.convertToImageBitmap(): ImageBitmap = this.asComposeImageBitmap()
